[server] collect types together
