--- UwuRadio.Server/Submitter.cs
@@ -1,3 +1,0 @@
-namespace UwuRadio.Server;
-
-public record Submitter(string Name, string PfpUrl, string[] Quotes);
